+++ packages/design-tokens/scripts/build-tokens.js
@@ -178,6 +178,7 @@ const colorModeFormat = require('../src/formats/color-mode-attributes')
     `tokens/functional/components/button/colors.js`,
     `tokens/functional/components/accordion/colors.js`,
     `tokens/functional/components/faq/colors.json`,
+    `tokens/functional/components/card/colors.json`,
     `tokens/functional/components/inline-link/colors.json`,
     `tokens/functional/components/control/colors.js`,
     `tokens/functional/components/subdomain-nav-bar/colors.js`,