+++ packages/react/src/Card/Card.module.css.d.ts
@@ -1,0 +1,8 @@
+declare const styles: {
+  readonly "Card": string;
+  readonly "Card__heading": string;
+  readonly "Card__description": string;
+  readonly "Card__action": string;
+};
+export = styles;
+